% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREPShiny.R
\name{PREPShiny}
\alias{PREPShiny}
\title{Launch 'PREPShiny' Interface}
\usage{
PREPShiny()
}
\value{
Nothing
}
\description{
PREPShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start PREPShiny

The interactive user interface is to provide an easy way for preprocessing the dataset and downloading the preprocessed dataset.
}
\examples{
if(interactive()){
library(rmarkdown)
PREPShiny()
}
}
\keyword{PREPShiny}
